+++ enterprise/frontend/src/metabase-enterprise/data_apps/components/DataAppIframeApp/DataAppIframeApp.tsx
@@ -1,14 +1,13 @@
+import { Center, Loader } from "frontend/src/metabase/ui";
+import { color } from "frontend/src/metabase/ui/colors";
 import { Component, type ErrorInfo, type ReactNode, useMemo } from "react";
 import { t } from "ttag";
 
-import { Center, Loader } from "metabase/ui";
-import { color } from "metabase/ui/colors";
-
-import { DataAppProvider } from "./components/DataAppProvider";
-import { type ErrorDetail, describeError } from "./lib/describe-error";
-import { readNameFromUrl } from "./lib/read-name-from-url";
-import { reportErrorToParent } from "./lib/report-error-to-parent";
-import { useDataAppBundle } from "./lib/use-data-app-bundle";
+import { type ErrorDetail, describeError } from "../../lib/describe-error";
+import { readNameFromUrl } from "../../lib/read-name-from-url";
+import { reportErrorToParent } from "../../lib/report-error-to-parent";
+import { useDataAppBundle } from "../../lib/use-data-app-bundle";
+import { DataAppProvider } from "../DataAppProvider/DataAppProvider";
 
 /**
  * Catches errors thrown while *rendering* the bundle's React tree (the async
