+++ enterprise/frontend/src/metabase-enterprise/data_apps/components/DataAppLayout/DataAppLayout.tsx
@@ -1,6 +1,5 @@
+import { Box } from "frontend/src/metabase/ui";
 import type { ReactNode } from "react";
-
-import { Box } from "metabase/ui";
 
 import S from "./DataAppLayout.module.css";
