+++ enterprise/frontend/src/metabase-enterprise/data_apps/components/DataAppLayout/DataAppLayout.unit.spec.tsx
@@ -1,4 +1,4 @@
-import { renderWithProviders, screen } from "__support__/ui";
+import { renderWithProviders, screen } from "frontend/test/__support__/ui";
 
 import { DataAppLayout } from "./DataAppLayout";
