+++ enterprise/frontend/src/metabase-enterprise/data_apps/components/DataAppProvider/DataAppProvider.tsx
@@ -1,15 +1,14 @@
 import createCache from "@emotion/cache";
 import { CacheProvider, Global } from "@emotion/react";
+import { SCOPED_CSS_RESET } from "frontend/src/embedding-sdk-bundle/components/private/PublicComponentStylesWrapper";
+import { PortalContainer } from "frontend/src/embedding-sdk-bundle/components/private/SdkPortalContainer";
+import { SdkThemeProvider } from "frontend/src/embedding-sdk-bundle/components/private/SdkThemeProvider";
+import { MetabaseReduxProvider } from "frontend/src/metabase/redux";
+import { getCspNonce } from "frontend/src/metabase/utils/csp";
 import { type ReactNode, useMemo } from "react";
 
-import { SCOPED_CSS_RESET } from "embedding-sdk-bundle/components/private/PublicComponentStylesWrapper";
-import { PortalContainer } from "embedding-sdk-bundle/components/private/SdkPortalContainer";
-import { SdkThemeProvider } from "embedding-sdk-bundle/components/private/SdkThemeProvider";
-import { MetabaseReduxProvider } from "metabase/redux";
-import { getCspNonce } from "metabase/utils/csp";
-
-import { useHostSdkStore } from "../lib/use-host-sdk-store";
-import type { DataAppMetabaseProviderProps } from "../sandbox";
+import { useHostSdkStore } from "../../lib/use-host-sdk-store";
+import type { DataAppMetabaseProviderProps } from "../../sandbox";
 
 // Note: Mantine + SDK CSS is loaded into the iframe via the `data-app-vendors`
 // Rspack entry (`<link rel="stylesheet">` in the iframe srcdoc). Don't side-
