+++ enterprise/frontend/src/metabase-enterprise/data_apps/components/DataAppView/DataAppView.unit.spec.tsx
@@ -1,13 +1,13 @@
 import { act } from "@testing-library/react";
+import { useGetDataAppQuery } from "enterprise/frontend/src/metabase-enterprise/api";
+import { createMockDataApp } from "frontend/src/metabase-types/api/mocks";
+import { renderWithProviders, screen } from "frontend/test/__support__/ui";
 
-import { renderWithProviders, screen } from "__support__/ui";
-import { useGetDataAppQuery } from "metabase-enterprise/api";
-import { createMockDataApp } from "metabase-types/api/mocks";
+import { DATA_APP_ERROR_MESSAGE_TYPE } from "../../constants";
 
-import { AppView } from "./AppView";
-import { DATA_APP_ERROR_MESSAGE_TYPE } from "./constants";
+import { DataAppView } from "./DataAppView";
 
-jest.mock("metabase-enterprise/api", () => ({
+jest.mock("enterprise/frontend/src/metabase-enterprise/api", () => ({
   ...jest.requireActual("metabase-enterprise/api"),
   useGetDataAppQuery: jest.fn(),
 }));
@@ -33,10 +33,10 @@ function setup({
     error,
   } as unknown as ReturnType<typeof useGetDataAppQuery>);
 
-  renderWithProviders(<AppView params={{ name }} />);
+  renderWithProviders(<DataAppView params={{ name }} />);
 }
 
-describe("AppView", () => {
+describe("DataAppView", () => {
   afterEach(() => jest.clearAllMocks());
 
   it("shows a not-found screen when the name is empty", () => {
