+++ enterprise/frontend/src/metabase-enterprise/data_apps/routes.tsx
@@ -1,8 +1,8 @@
 import { IndexRoute, Route } from "metabase/router";
 import * as Urls from "metabase/urls";
 
-import { AppView } from "./AppView";
-import { DataAppLayout } from "./DataAppLayout";
+import { DataAppLayout } from "./components/DataAppLayout/DataAppLayout";
+import { DataAppView } from "./components/DataAppView/DataAppView";
 
 /**
  * Data-app host routes. Open to any signed-in user.
@@ -16,12 +16,12 @@ export function getRoutes() {
       path={`${Urls.DATA_APP_URL_SEGMENT}/:name`}
       component={DataAppLayout}
     >
-      <IndexRoute component={AppView} />
+      <IndexRoute component={DataAppView} />
       {/* Sub-paths under /apps/:name are owned by the iframe's router.
           Same component — `AppView` just keeps the iframe mounted; the URL
           change is mirrored back from inside the iframe via
           `history.replaceState`. */}
-      <Route path="*" component={AppView} />
+      <Route path="*" component={DataAppView} />
     </Route>
   );
 }
